requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy~=2.0.36
pydantic==2.10.0
pydantic-settings==2.6.1
alembic==1.14.0
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-cov==6.0.0
python-dotenv==1.0.1
flake8==7.1.1
black==24.10.0
httpx==0.27.2
Faker==33.0.0
